style: Input 파일 정리
