+++ src/components/common/ImageBox/ImageBoxLayout.tsx
@@ -6,13 +6,14 @@ interface ImageBox {
   size: string;
   src: string;
   alt: string;
+  disabled?: boolean;
 }
 
 const cx = classNames.bind(styles);
 
-export default function ImageBoxLayout({ size, src, alt }: ImageBox) {
+export default function ImageBoxLayout({ size, src, alt, disabled, ...rest }: ImageBox) {
   return (
-    <div className={cx(size, 'imageBox')}>
+    <div className={cx(size, 'imageBox', { disabled })} {...rest}>
       <Image src={src} alt={alt} fill />
     </div>
   );
