+++ src/components/common/Input/Input.module.scss
@@ -26,7 +26,7 @@
   }
 
   .label {
-    font-size: 0.875rem;
+    font-size: 2rem;
   }
 
   .inputWithLabel {
