+++ src/components/common/Input/index.tsx
@@ -1,50 +1,99 @@
+import { forwardRef, ChangeEvent } from 'react';
+import Image, { ImageProps } from 'next/image';
 import classNames from 'classnames/bind';
-import InputLayout from './InputLayout';
 import styles from './Input.module.scss';
+
+interface InputProps {
+  id: string;
+  type: string;
+  label?: string;
+  size?: string;
+  border?: string;
+  isError?: boolean;
+  errorText?: string;
+  labelStyle?: string;
+  placeholder: string;
+  imageProps?: Partial<ImageProps>;
+  background?: string;
+}
 
 const cx = classNames.bind(styles);
 
-export default function Input() {
+const Input = forwardRef<HTMLInputElement, InputProps>(function Input(
+  { id, label, isError, errorText, labelStyle, size, border, imageProps, background, ...rest },
+  ref
+) {
+  const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
+    if (ref && 'current' in ref && ref.current) {
+      ref.current.value = event.target.value;
+    }
+  };
   return (
-    <>
-      <InputLayout id="" type="text" size={cx('')} label=" " labelStyle={cx('')} placeholder="" />
-    </>
+    <div className={cx('inputWithLabel')}>
+      {label && (
+        <label htmlFor={id} className={cx(labelStyle)}>
+          {label}
+        </label>
+      )}
+      <div className={cx('inputWithIcon', size)}>
+        <input
+          ref={ref}
+          className={cx(border, { error: isError }, size, background)}
+          onChange={handleChange}
+          {...rest}
+        />
+        {imageProps && (
+          <Image
+            src="/images/search.svg"
+            alt="검색 아이콘"
+            width={24}
+            height={24}
+            className={cx('icon')}
+            {...imageProps}
+          />
+        )}
+      </div>
+      {isError && <p className={cx('errorText')}>{errorText}</p>}
+    </div>
   );
-}
+});
+
+export default Input;
 
 {
   /* 사용법
+  기본: <Input id="" type="text" size={''} label=" " labelStyle="label" placeholder="" background={''} />;
   1. 닉네임 인풋
-      <InputLayout
+      <Input
         id=""
         type="text"
-        size={cx('large')}
+        size={'large'}
         label=" "
-        labelStyle={cx('label')}
+        labelStyle={'label'}
         placeholder="닉네임을 입력해주세요"
       />
   2.  사이즈 큰 검색 인풋
-      <InputLayout
+      <Input
         id=""
         type="text"
-        size={cx('mediumLarge')}
-        border={cx('roundBorder')}
+        size={'mediumLarge'}
+        border={'roundBorder'}
         label=" "
-        labelStyle={cx('label')}
+        labelStyle={'label'}
         placeholder="검색어를 입력해주세요"
         imageProps={{}}
-        background={cx('background')}
+        background={'background'}
       />
   3. 사이즈 작은 검색 인풋
-      <InputLayout
+      <Input
         id=""
         type="text"
-        size={cx('medium')}
-        border={cx('roundBorder')}
+        size={'medium'}
+        border={'roundBorder'}
         label=" "
-        labelStyle={cx('label')}
+        labelStyle={'label'}
         placeholder="검색어를 입력해주세요"
         imageProps={{}}
-        background={cx('background')}
+        background={'background'}
       /> */
 }
